# ms-carrito/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/carrito": {
            "get": {
                "description": "get carritos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Devuelve todos los carritos de la base de datos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CarritoFindAllResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/carrito/cursos": {
            "get": {
                "description": "get cursos para cada carrito",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Devuelve para todos los carritos, los cursos que contiene",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CursoCarritoFindAllResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/carrito/{id}": {
            "get": {
                "description": "Busca un carrito por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Devuelve por ID un único carrito",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CARRITO ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CarritoFindbyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{cliente_id}/carrito": {
            "post": {
                "description": "Agrega un carrito a la base de datos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Crea un carrito",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id cliente",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CarritoCreateResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un carrito a partir del ID del cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Elimina un carrito",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del cliente",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CarritoUpdateResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{cliente_id}/carrito/curso": {
            "get": {
                "description": "Busca cursos por cliente id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Devuelve los cursos que tiene un carrito por cliente ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "CLIENTE ID",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CursoCarritoFindbyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{cliente_id}/carrito/curso/all": {
            "get": {
                "description": "cantidad cursos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cliente"
                ],
                "summary": "catnid cursos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cliente ID",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetCantidadCursos"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina todos los cursos de un cliente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cliente"
                ],
                "summary": "Elimina todos los cursos a partir del id cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cliente ID",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CursoCarritoDeleteAllResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{cliente_id}/carrito/curso/{curso_id}": {
            "post": {
                "description": "Agrega un curso a un carrito mediante body",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Agrega un curso a un carrito",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cliente ID",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Curso ID",
                        "name": "curso_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CursoCarritoCreateResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Para un carrito se eliminar un curso",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carritos"
                ],
                "summary": "Elemina un curso de un carrito",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cliente ID",
                        "name": "cliente_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Curso ID",
                        "name": "curso_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CursoCarritoDeleteOneResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CreateCursoCarritoRequest": {
            "type": "object",
            "required": [
                "carrito_id",
                "curso_id"
            ],
            "properties": {
                "carrito_id": {
                    "type": "integer"
                },
                "curso_id": {
                    "type": "integer"
                }
            }
        },
        "request.DeleteCursoCarritoRequest": {
            "type": "object",
            "required": [
                "carrito_id",
                "curso_id"
            ],
            "properties": {
                "carrito_id": {
                    "type": "integer"
                },
                "curso_id": {
                    "type": "integer"
                }
            }
        },
        "response.CarritoCreateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CarritoFindAllResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CarritoResponse"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CarritoFindbyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/response.CarritoResponse"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CarritoResponse": {
            "type": "object",
            "properties": {
                "carrito_id": {
                    "type": "integer"
                },
                "cliente_id": {
                    "type": "integer"
                }
            }
        },
        "response.CarritoUpdateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoCreateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/request.CreateCursoCarritoRequest"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoDeleteAllResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoDeleteOneResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/request.DeleteCursoCarritoRequest"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoFindAllResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CursoCarritoResponse"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoFindbyResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.CursoResponse"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.CursoCarritoResponse": {
            "type": "object",
            "properties": {
                "carrito_id": {
                    "type": "integer"
                },
                "curso_id": {
                    "type": "integer"
                }
            }
        },
        "response.CursoResponse": {
            "type": "object",
            "properties": {
                "curso_id": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.GetCantidadCursos": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}